feat(FE): #53 코드 리팩토링 - gql, gql request, response type 관리 포인트 분리위해 파일 추가, 기존 type/index.ts -> store.ts (클라이언트 상태관리), api.ts (api 관련 요청, 응답값) 로 분리, 이 다음에 response 받아서 처리하는 부분 onComplete, onError로 처리 수정 필요
