+++ frontend/src/pages/Notice.tsx
@@ -1,55 +1,12 @@
-import { gql, useQuery } from "@apollo/client";
+import { useQuery } from "@apollo/client";
 import { useEffect } from "react";
 import { useResetRecoilState, useRecoilState } from "recoil";
+import { GET_NOTICES } from "../gql";
 import ListLayout from "../layouts/Admin/ListLayout";
 import { checkedListState, pageInfoState } from "../recoil";
-import { ListItem } from "../types";
+import { NoticeListResponse, NoticeListVariable } from "../types/api";
+import { ListItem } from "../types/store";
 import { getNoticeHeaderList } from "../util";
-
-interface NoticeVariable {
-  pagination: {
-    skip: number;
-    take: number;
-  };
-  orderBy: {
-    createdAt: string;
-  };
-}
-
-interface NoticeItem {
-  authorId: string;
-  content: string;
-  createdAt: number;
-  postId: string;
-  status: string;
-  title: string;
-  updatedAt: number;
-}
-
-interface NoticeResponse {
-  posts: { list: NoticeItem[]; totalCount: number };
-}
-
-const GET_POSTS = gql`
-  query GET_POSTS($pagination: OffsetPagination!, $orderBy: PostsOrder) {
-    posts(pagination: $pagination, orderBy: $orderBy) {
-      list {
-        postId
-        title
-        content
-        status
-        authorId
-        author {
-          id
-          userName
-        }
-        updatedAt
-        createdAt
-      }
-      totalCount
-    }
-  }
-`;
 
 const Notice = () => {
   const [pageInfo, setPageInfo] = useRecoilState(pageInfoState);
@@ -60,20 +17,20 @@ const Notice = () => {
   } = getNoticeHeaderList();
 
   let list: ListItem[] = [];
-  const { loading, error, data } = useQuery<NoticeResponse, NoticeVariable>(
-    GET_POSTS,
-    {
-      variables: {
-        pagination: {
-          skip: pageInfo.offset * (pageInfo.page - 1),
-          take: pageInfo.offset,
-        },
-        orderBy: {
-          createdAt: "desc",
-        },
+  const { loading, error, data } = useQuery<
+    NoticeListResponse,
+    NoticeListVariable
+  >(GET_NOTICES, {
+    variables: {
+      pagination: {
+        skip: pageInfo.offset * (pageInfo.page - 1),
+        take: pageInfo.offset,
       },
-    }
-  );
+      orderBy: {
+        createdAt: "desc",
+      },
+    },
+  });
 
   if (loading || error) {
     list = [];
