+++ frontend/src/pages/Notice/NoticeUpdate.tsx
@@ -1,90 +1,17 @@
 import { MouseEvent, useEffect } from "react";
 import { useRecoilState } from "recoil";
-import { gql, useMutation, useQuery } from "@apollo/client";
+import { useMutation, useQuery } from "@apollo/client";
 import { useToast } from "@chakra-ui/react";
 import { useParams, useNavigate } from "react-router-dom";
 import UpdateLayout from "../../layouts/Admin/UpdateLayout";
 import { noticeItemState } from "../../recoil";
-
-interface PostVariable {
-  noticeId: string;
-}
-
-interface PostResponse {
-  getPostById: {
-    postId: string;
-    title: string;
-    content: string;
-    status: string;
-    authorId: string;
-    author: {
-      id: string;
-      userName: string;
-    };
-    updatedAt: number;
-    createdAt: number;
-  };
-}
-
-const GET_NOTICE = gql`
-  query GetPostById($noticeId: String!) {
-    getPostById(id: $noticeId) {
-      postId
-      title
-      content
-      status
-      authorId
-      author {
-        id
-        userName
-      }
-      updatedAt
-      createdAt
-    }
-  }
-`;
-
-const CREATE_NOTICE = gql`
-  mutation Mutation($notice: CreatePostInput!) {
-    createPost(post: $notice) {
-      postId
-      title
-      content
-      status
-      authorId
-      updatedAt
-      createdAt
-    }
-  }
-`;
-
-const UPDATE_NOTICE = gql`
-  mutation Mutation($notice: UpdatePostInput!) {
-    updatePost(post: $notice) {
-      postId
-      title
-      content
-      status
-      authorId
-      updatedAt
-      createdAt
-    }
-  }
-`;
-
-const DELETE_NOTICE = gql`
-  mutation Mutation($notice: DeletePostInput!) {
-    deletePost(id: $notice) {
-      postId
-      title
-      content
-      status
-      authorId
-      updatedAt
-      createdAt
-    }
-  }
-`;
+import {
+  CREATE_NOTICE,
+  DELETE_NOTICE,
+  GET_NOTICE,
+  UPDATE_NOTICE,
+} from "../../gql";
+import { NoticeResponse, NoticeVariable } from "../../types/api";
 
 const NoticeUpdate = () => {
   const params = useParams();
@@ -93,18 +20,20 @@ const NoticeUpdate = () => {
   const [notice, setNotice] = useRecoilState(noticeItemState);
   const noticeId = (params && params.id) || "";
 
-  const { loading, data } = useQuery<PostResponse, PostVariable>(GET_NOTICE, {
-    variables: {
-      noticeId,
-    },
-  });
+  const { loading, data } = useQuery<NoticeResponse, NoticeVariable>(
+    GET_NOTICE,
+    {
+      variables: {
+        noticeId,
+      },
+    }
+  );
 
   // TODO: 저장과 발행의 차이는 무엇인지 확인하기
   const [updateNotice, { error: updateError, data: updateData }] = useMutation(
     noticeId ? UPDATE_NOTICE : CREATE_NOTICE,
     {
       onCompleted: () => {
-        console.log("onCompleted");
         const type = "save";
         const success = updateData && !updateError;
         toast({
@@ -117,7 +46,6 @@ const NoticeUpdate = () => {
         });
       },
       onError: () => {
-        console.log("onError");
         const type = "save";
         const success = updateData && !updateError;
         toast({
@@ -136,7 +64,6 @@ const NoticeUpdate = () => {
     DELETE_NOTICE,
     {
       onCompleted: () => {
-        console.log("delete>onCompleted");
         const success = deleteData && deleteData.deletePost && !deleteError;
         toast({
           description: `삭제를 ${success === 0 ? "완료" : "실패"}했습니다.`,
@@ -146,7 +73,6 @@ const NoticeUpdate = () => {
         });
       },
       onError: () => {
-        console.log("delete>onError");
         const success = deleteData && deleteData.deletePost && !deleteError;
         toast({
           description: `삭제를 ${success === 0 ? "완료" : "실패"}했습니다.`,
